requirements.txt
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
blis==0.4.1
cachetools==3.1.1
catalogue==0.0.8
certifi==2019.9.11
chardet==3.0.4
coverage==4.5.4
cycler==0.10.0
cymem==2.0.3
cytoolz==0.10.1
dask==2.8.1
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
h5py==2.10.0
hypothesis==4.47.1
idna==2.8
importlib-metadata==0.23
jedi==0.15.1
jellyfish==0.7.2
Jinja2==2.10.3
joblib==0.14.0
jsonschema==3.2.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==7.2.0
murmurhash==1.0.2
nbformat==4.4.0
networkx==2.4
nose==1.3.7
numpy==1.17.4
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
py==1.8.0
pyemd==0.5.1
Pygments==2.4.2
pyparsing==2.4.5
Pyphen==0.9.5
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.1
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.3
Send2Trash==1.5.0
six==1.13.0
spacy==2.2.3
srsly==0.2.0
terminado==0.8.3
testpath==0.4.4
textacy==0.9.1
thinc==7.3.1
toolz==0.10.0
tornado==6.0.3
tqdm==4.39.0
traitlets==4.3.3
urllib3==1.25.7
wasabi==0.4.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==0.6.0
